requirements.txt
fastapi==0.115.12
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.59
langchain_google_genai==2.1.4
langchain_groq==0.3.2
langchain_huggingface==0.2.0
motor==3.7.0
pandas==2.2.3
passlib==1.7.4
pydantic==2.11.4
pymongo==4.12.1
pypdf==5.5.0
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.44.1
uvicorn==0.34.2
